+++ nio/api.py
@@ -29,11 +29,15 @@ from __future__ import unicode_literals
 import json
 from collections import defaultdict
 from enum import Enum, unique
-from typing import (Any, DefaultDict, Dict, Iterable, List,
-                    Optional, Set, Sequence, Tuple, Union)
+from typing import (
+    Any, DefaultDict, Dict, Iterable, List, Optional, Sequence, Set, Tuple,
+    Union, TYPE_CHECKING,
+)
 
 from .exceptions import LocalProtocolError
-from .http import Http2Request, HttpRequest, TransportRequest
+
+if TYPE_CHECKING:
+    from .events.account_data import PushAction, PushCondition
 
 if False:
     from uuid import UUID
@@ -111,6 +115,17 @@ class EventFormat(Enum):
 
     client = "client"
     federation = "federation"
+
+
+@unique
+class PushRuleKind(Enum):
+    """Push rule kinds defined by the Matrix spec, ordered by priority."""
+
+    override = "override"
+    content = "content"
+    room = "room"
+    sender = "sender"
+    underride = "underride"
 
 
 class Api:
@@ -1471,19 +1486,28 @@ class Api:
         )
 
     @staticmethod
-    def profile_get(user_id: str) -> Tuple[str, str]:
+    def profile_get(user_id: str, access_token: str = None) -> Tuple[str, str]:
         """Get the combined profile information for a user.
 
         Returns the HTTP method and HTTP path for the request.
 
         Args:
             user_id (str): User id to get the profile for.
+            access_token (str): The access token to be used with the request. If
+                                omitted, an unauthenticated request is perfomed.
         """
         assert user_id
-        return "GET", Api._build_path(path=["profile", user_id])
+
+        query_parameters = {}
+        if access_token is not None:
+            query_parameters["access_token"] = access_token
+
+        path = ["profile", user_id]
+
+        return "GET", Api._build_path(path, query_parameters)
 
     @staticmethod
-    def profile_get_displayname(user_id):
+    def profile_get_displayname(user_id, access_token: str = None):
         # type (str, str) -> Tuple[str, str]
         """Get display name.
 
@@ -1491,10 +1515,16 @@ class Api:
 
         Args:
             user_id (str): User id to get display name for.
+            access_token (str): The access token to be used with the request. If
+                                omitted, an unauthenticated request is perfomed.
         """
+        query_parameters = {}
+        if access_token is not None:
+            query_parameters["access_token"] = access_token
+
         path = ["profile", user_id, "displayname"]
 
-        return "GET", Api._build_path(path)
+        return "GET", Api._build_path(path, query_parameters)
 
     @staticmethod
     def profile_set_displayname(access_token, user_id, display_name):
@@ -1519,7 +1549,7 @@ class Api:
         )
 
     @staticmethod
-    def profile_get_avatar(user_id):
+    def profile_get_avatar(user_id, access_token: str = None):
         # type (str, str) -> Tuple[str, str]
         """Get avatar URL.
 
@@ -1527,10 +1557,15 @@ class Api:
 
         Args:
             user_id (str): User id to get avatar for.
+            access_token (str): The access token to be used with the request. If
+                                omitted, an unauthenticated request is perfomed.
         """
+        query_parameters = {}
+        if access_token is not None:
+            query_parameters["access_token"] = access_token
         path = ["profile", user_id, "avatar_url"]
 
-        return "GET", Api._build_path(path)
+        return "GET", Api._build_path(path, query_parameters)
 
     @staticmethod
     def profile_set_avatar(access_token, user_id, avatar_url):
@@ -1689,6 +1724,183 @@ class Api:
 
         return (
             "POST",
+            Api._build_path(path, query_parameters),
+            Api.to_json(content),
+        )
+
+    @staticmethod
+    def set_pushrule(
+        access_token: str,
+        scope: str,
+        kind: PushRuleKind,
+        rule_id: str,
+        before: Optional[str] = None,
+        after: Optional[str] = None,
+        actions: Sequence["PushAction"] = (),
+        conditions: Optional[Sequence["PushCondition"]] = None,
+        pattern: Optional[str] = None,
+    ) -> Tuple[str, str, str]:
+        """Create or modify an existing user-created push rule.
+
+        Returns the HTTP method, HTTP path and data for the request.
+
+        Args:
+            access_token (str): The access token to be used with the request.
+
+            scope (str): The scope of this rule, e.g. ``"global"``.
+                Homeservers currently only process ``global`` rules for
+                event matching, while ``device`` rules are a planned feature.
+                It is up to clients to interpret any other scope name.
+
+            kind (PushRuleKind): The kind of rule.
+
+            rule_id (str): The identifier of the rule. Must be unique
+                within its scope and kind.
+                For rules of ``room`` kind, this is the room ID to match for.
+                For rules of ``sender`` kind, this is the user ID to match.
+
+            before (Optional[str]): Position this rule before the one matching
+                the given rule ID.
+                The rule ID cannot belong to a predefined server rule.
+                ``before`` and ``after`` cannot be both specified.
+
+            after (Optional[str]): Position this rule after the one matching
+                the given rule ID.
+                The rule ID cannot belong to a predefined server rule.
+                ``before`` and ``after`` cannot be both specified.
+
+            actions (Sequence[PushAction]): Actions to perform when the
+                conditions for this rule are met. The given actions replace
+                the existing ones.
+
+            conditions (Sequence[PushCondition]): Event conditions that must
+                hold true for the rule to apply to that event.
+                A rule with no conditions always hold true.
+                Only applicable to ``underride`` and ``override`` rules.
+
+            pattern (Optional[str]): Glob-style pattern to match against
+                for the event's content.
+                Only applicable to ``content`` rules.
+        """
+
+        path = ["pushrules", scope, kind.value, rule_id]
+        query_parameters = {"access_token": access_token}
+        content: Dict[str, Any] = {"actions": [a.as_value for a in actions]}
+
+        if before is not None and after is not None:
+            raise TypeError("before and after cannot be both specified")
+        elif before is not None:
+            query_parameters["before"] = before
+        elif after is not None:
+            query_parameters["after"] = after
+
+        if pattern is not None:
+            if kind != PushRuleKind.content:
+                raise TypeError("pattern can only be set for content rules")
+
+            content["pattern"] = pattern
+
+        if conditions is not None:
+            if kind not in (PushRuleKind.override, PushRuleKind.underride):
+                raise TypeError(
+                    "conditions can only be set for override/underride rules",
+                )
+
+            content["conditions"] = [c.as_value for c in conditions]
+
+        return (
+            "PUT",
+            Api._build_path(path, query_parameters),
+            Api.to_json(content),
+        )
+
+    @staticmethod
+    def delete_pushrule(
+        access_token: str, scope: str, kind: PushRuleKind, rule_id: str,
+    ) -> Tuple[str, str]:
+        """Delete an existing user-created push rule.
+
+        Returns the HTTP method and HTTP path for the request.
+
+        Args:
+            access_token (str): The access token to be used with the request.
+            scope (str): The scope of this rule, e.g. ``"global"``.
+            kind (PushRuleKind): The kind of rule.
+            rule_id (str): The identifier of the rule. Must be unique
+                within its scope and kind.
+        """
+
+        path = ["pushrules", scope, kind.value, rule_id]
+        query_parameters = {"access_token": access_token}
+
+        return ("DELETE", Api._build_path(path, query_parameters))
+
+    @staticmethod
+    def enable_pushrule(
+        access_token: str,
+        scope: str,
+        kind: PushRuleKind,
+        rule_id: str,
+        enable: bool,
+    ) -> Tuple[str, str, str]:
+        """Enable or disable an existing built-in or user-created push rule.
+
+        Returns the HTTP method, HTTP path and data for the request.
+
+        Args:
+            access_token (str): The access token to be used with the request.
+            scope (str): The scope of this rule, e.g. ``"global"``.
+            kind (PushRuleKind): The kind of rule.
+            rule_id (str): The identifier of the rule. Must be unique
+                within its scope and kind.
+            enable (bool): Whether to enable or disable the rule.
+        """
+
+        path = ["pushrules", scope, kind.value, rule_id, "enabled"]
+        query_parameters = {"access_token": access_token}
+        content = {"enabled": enable}
+
+        return (
+            "PUT",
+            Api._build_path(path, query_parameters),
+            Api.to_json(content),
+        )
+
+    @staticmethod
+    def set_pushrule_actions(
+        access_token: str,
+        scope: str,
+        kind: PushRuleKind,
+        rule_id: str,
+        actions: Sequence["PushAction"],
+    ) -> Tuple[str, str, str]:
+        """Set the actions for an existing built-in or user-created push rule.
+
+        Unlike ``set_pushrule``, this method can edit built-in server rules.
+
+        Returns the HTTP method, HTTP path and data for the request.
+
+        Args:
+            access_token (str): The access token to be used with the request.
+
+            scope (str): The scope of this rule, e.g. ``"global"``.
+
+            kind (PushRuleKind): The kind of rule.
+
+            rule_id (str): The identifier of the rule. Must be unique
+                within its scope and kind.
+
+            actions (Sequence[PushAction]): Actions to perform when the
+                conditions for this rule are met. The given actions replace
+                the existing ones.
+        """
+
+        path = ["pushrules", scope, kind.value, rule_id, "actions"]
+        query_parameters = {"access_token": access_token}
+        content = {"actions": [a.as_value for a in actions]}
+
+        return (
+            "PUT",
             Api._build_path(path, query_parameters),
             Api.to_json(content),
         )